+++ LearningAPI/apps.py
@@ -1,5 +1,8 @@
 from django.apps import AppConfig
 
-
-class LearningapiConfig(AppConfig):
+class LearningAPIConfig(AppConfig):
     name = 'LearningAPI'
+
+    def ready(self):
+        # Import the signals module to ensure signal handlers are connected
+        import LearningAPI.signals
